Add test cases for struct collection: - Serializing a default instance. - Deserializing a null json string.
