+++ src/libraries/System.Text.Json/tests/Serialization/CollectionTests/CollectionTests.Dictionary.cs
@@ -278,6 +278,21 @@ namespace System.Text.Json.Serialization.Tests
             {
                 GenericStructIDictionaryWrapper<string, string>? obj = JsonSerializer.Deserialize<GenericStructIDictionaryWrapper<string, string>?>("null");
                 Assert.False(obj.HasValue);
+
+                string json = JsonSerializer.Serialize(obj);
+                Assert.Equal("null", json);
+            }
+
+            {
+                GenericStructIDictionaryWrapper<string, string> obj = default;
+                string json = JsonSerializer.Serialize(obj);
+                Assert.Equal("{}", json);
+            }
+
+            {
+                StructWrapperForIDictionary obj = default;
+                string json = JsonSerializer.Serialize(obj);
+                Assert.Equal("{}", json);
             }
         }
