+++ src/libraries/System.Text.Json/tests/Serialization/TestClasses/TestClasses.GenericCollections.cs
@@ -1174,4 +1174,471 @@ namespace System.Text.Json.Serialization.Tests
     public interface IDerivedIDictionaryOfTKeyTValue<TKey, TValue> : IDictionary<TKey, TValue> { }
 
     public interface IDerivedISetOfT<T> : ISet<T> { }
+
+    public struct GenericStructIListWrapper<T> : IList<T>
+    {
+        private List<T> _list;
+        public T this[int index]
+        {
+            get
+            {
+                InitializeIfNull();
+                return _list[index];
+            }
+            set
+            {
+                InitializeIfNull();
+                _list[index] = value;
+            }
+        }
+
+        public int Count => _list == null ? 0 : _list.Count;
+
+        public bool IsReadOnly => false;
+
+        private void InitializeIfNull()
+        {
+            if (_list == null)
+            {
+                _list = new List<T>();
+            }
+        }
+
+        public void Add(T item)
+        {
+            InitializeIfNull();
+            _list.Add(item);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            _list.Clear();
+        }
+
+        public bool Contains(T item)
+        {
+            InitializeIfNull();
+            return _list.Contains(item);
+        }
+
+        public void CopyTo(T[] array, int arrayIndex)
+        {
+            InitializeIfNull();
+            throw new NotImplementedException();
+        }
+
+        public IEnumerator<T> GetEnumerator()
+        {
+            InitializeIfNull();
+            return _list.GetEnumerator();
+        }
+
+        public int IndexOf(T item)
+        {
+            InitializeIfNull();
+            return _list.IndexOf(item);
+        }
+
+        public void Insert(int index, T item)
+        {
+            InitializeIfNull();
+            _list.Insert(index, item);
+        }
+
+        public bool Remove(T item)
+        {
+            InitializeIfNull();
+            return _list.Remove(item);
+        }
+
+        public void RemoveAt(int index)
+        {
+            InitializeIfNull();
+            _list.RemoveAt(index);
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            InitializeIfNull();
+            return GetEnumerator();
+        }
+    }
+
+    public struct GenericStructICollectionWrapper<T> : ICollection<T>
+    {
+        private List<T> _list;
+
+        private void InitializeIfNull()
+        {
+            if (_list == null)
+            {
+                _list = new List<T>();
+            }
+        }
+
+        public int Count => _list.Count;
+
+        public bool IsReadOnly => false;
+
+        public void Add(T item)
+        {
+            InitializeIfNull();
+            _list.Add(item);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            _list.Clear();
+        }
+
+        public bool Contains(T item)
+        {
+            InitializeIfNull();
+            return _list.Contains(item);
+        }
+
+        public void CopyTo(T[] array, int arrayIndex)
+        {
+            InitializeIfNull();
+            _list.CopyTo(array, arrayIndex);
+        }
+
+        public IEnumerator<T> GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((ICollection<T>)_list).GetEnumerator();
+        }
+
+        public bool Remove(T item)
+        {
+            InitializeIfNull();
+            return _list.Remove(item);
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((ICollection<T>)_list).GetEnumerator();
+        }
+    }
+
+    public struct GenericStructISetWrapper<T> : ISet<T>
+    {
+        private HashSet<T> _hashset;
+
+        private void InitializeIfNull()
+        {
+            if (_hashset == null)
+            {
+                _hashset = new HashSet<T>();
+            }
+        }
+
+        public int Count => _hashset == null ? 0 : _hashset.Count;
+
+        public bool IsReadOnly => false;
+
+        public bool Add(T item)
+        {
+            InitializeIfNull();
+            return _hashset.Add(item);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            _hashset.Clear();
+        }
+
+        public bool Contains(T item)
+        {
+            InitializeIfNull();
+            return _hashset.Contains(item);
+        }
+
+        public void CopyTo(T[] array, int arrayIndex)
+        {
+            InitializeIfNull();
+            _hashset.CopyTo(array, arrayIndex);
+        }
+
+        public void ExceptWith(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            _hashset.ExceptWith(other);
+        }
+
+        public IEnumerator<T> GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((ISet<T>)_hashset).GetEnumerator();
+        }
+
+        public void IntersectWith(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            _hashset.IntersectWith(other);
+        }
+
+        public bool IsProperSubsetOf(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.IsProperSubsetOf(other);
+        }
+
+        public bool IsProperSupersetOf(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.IsProperSupersetOf(other);
+        }
+
+        public bool IsSubsetOf(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.IsSubsetOf(other);
+        }
+
+        public bool IsSupersetOf(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.IsSupersetOf(other);
+        }
+
+        public bool Overlaps(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.Overlaps(other);
+        }
+
+        public bool Remove(T item)
+        {
+            InitializeIfNull();
+            return _hashset.Remove(item);
+        }
+
+        public bool SetEquals(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            return _hashset.SetEquals(other);
+        }
+
+        public void SymmetricExceptWith(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            _hashset.SymmetricExceptWith(other);
+        }
+
+        public void UnionWith(IEnumerable<T> other)
+        {
+            InitializeIfNull();
+            _hashset.UnionWith(other);
+        }
+
+        void ICollection<T>.Add(T item)
+        {
+            InitializeIfNull();
+            _hashset.Add(item);
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((ISet<T>)_hashset).GetEnumerator();
+        }
+    }
+
+    public struct GenericStructIDictionaryWrapper<TKey, TValue> : IDictionary<TKey, TValue>
+    {
+        private Dictionary<TKey, TValue> _dict;
+
+        private void InitializeIfNull()
+        {
+            if (_dict == null)
+            {
+                _dict = new Dictionary<TKey, TValue>();
+            }
+        }
+
+        public TValue this[TKey key]
+        {
+            get
+            {
+                InitializeIfNull();
+                return _dict[key];
+            }
+            set
+            {
+                InitializeIfNull();
+                _dict[key] = value;
+            }
+        }
+
+        public ICollection<TKey> Keys => ((IDictionary<TKey, TValue>)_dict).Keys;
+
+        public ICollection<TValue> Values => ((IDictionary<TKey, TValue>)_dict).Values;
+
+        public int Count => _dict == null ? 0 : _dict.Count;
+
+        public bool IsReadOnly => false;
+
+        public void Add(TKey key, TValue value)
+        {
+            InitializeIfNull();
+            ((IDictionary<TKey, TValue>)_dict).Add(key, value);
+        }
+
+        public void Add(KeyValuePair<TKey, TValue> item)
+        {
+            InitializeIfNull();
+            ((IDictionary<TKey, TValue>)_dict).Add(item);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            ((IDictionary<TKey, TValue>)_dict).Clear();
+        }
+
+        public bool Contains(KeyValuePair<TKey, TValue> item)
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).Contains(item);
+        }
+
+        public bool ContainsKey(TKey key)
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).ContainsKey(key);
+        }
+
+        public void CopyTo(KeyValuePair<TKey, TValue>[] array, int arrayIndex)
+        {
+            InitializeIfNull();
+            ((IDictionary<TKey, TValue>)_dict).CopyTo(array, arrayIndex);
+        }
+
+        public IEnumerator<KeyValuePair<TKey, TValue>> GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).GetEnumerator();
+        }
+
+        public bool Remove(TKey key)
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).Remove(key);
+        }
+
+        public bool Remove(KeyValuePair<TKey, TValue> item)
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).Remove(item);
+        }
+
+        public bool TryGetValue(TKey key, out TValue value)
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).TryGetValue(key, out value);
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((IDictionary<TKey, TValue>)_dict).GetEnumerator();
+        }
+    }
+
+    public class ClassWithGenericStructIListWrapper
+    {
+        public GenericStructIListWrapper<int> List { get; set; }
+    }
+
+    public class ClassWithGenericStructICollectionWrapper
+    {
+        public GenericStructICollectionWrapper<int> Collection { get; set; }
+    }
+
+    public class ClassWithGenericStructIDictionaryWrapper
+    {
+        public GenericStructIDictionaryWrapper<string, string> Dictionary { get; set; }
+    }
+
+    public class ClassWithGenericStructISetWrapper
+    {
+        public GenericStructISetWrapper<int> Set { get; set; }
+    }
+
+    public class SimpleTestClassWithGenericStructCollectionWrappers : ITestClass
+    {
+        public GenericStructIListWrapper<int> List { get; set; }
+        public GenericStructICollectionWrapper<int> Collection { get; set; }
+        public GenericStructISetWrapper<int> Set { get; set; }
+        public GenericStructIDictionaryWrapper<string, string> Dictionary { get; set; }
+
+        public static readonly string s_json =
+            @"{" +
+            @"""List"" : [10]," +
+            @"""Collection"" : [30]," +
+            @"""Set"" : [50]," +
+            @"""Dictionary"" : {""key1"" : ""value1""}" +
+            @"}";
+
+        public void Initialize()
+        {
+            List = new GenericStructIListWrapper<int>() { 10 };
+            Collection = new GenericStructICollectionWrapper<int>() { 30 };
+            Set = new GenericStructISetWrapper<int>() { 50 };
+            Dictionary = new GenericStructIDictionaryWrapper<string, string>() { { "key1", "value1" } };
+        }
+
+        public void Verify()
+        {
+            Assert.Equal(1, List.Count);
+            Assert.Equal(10, List[0]);
+            Assert.Equal(1, Collection.Count);
+            Assert.Equal(30, Collection.ElementAt(0));
+            Assert.Equal(1, Set.Count);
+            Assert.Equal(50, Set.ElementAt(0));
+            Assert.Equal(1, Dictionary.Keys.Count);
+            Assert.Equal("value1", Dictionary["key1"]);
+        }
+    }
+
+    public struct SimpleTestStructWithNullableGenericStructCollectionWrappers : ITestClass
+    {
+        public GenericStructIListWrapper<int>? List { get; set; }
+        public GenericStructICollectionWrapper<int>? Collection { get; set; }
+        public GenericStructISetWrapper<int>? Set { get; set; }
+        public GenericStructIDictionaryWrapper<string, string>? Dictionary { get; set; }
+
+        public static readonly string s_json =
+            @"{" +
+            @"""List"" : [10]," +
+            @"""Collection"" : [30]," +
+            @"""Set"" : [50]," +
+            @"""Dictionary"" : {""key1"" : ""value1""}" +
+            @"}";
+
+        public void Initialize()
+        {
+            List = new GenericStructIListWrapper<int>() { 10 };
+            Collection = new GenericStructICollectionWrapper<int>() { 30 };
+            Set = new GenericStructISetWrapper<int>() { 50 };
+            Dictionary = new GenericStructIDictionaryWrapper<string, string>() { { "key1", "value1" } };
+        }
+
+        public void Verify()
+        {
+            Assert.Equal(1, List.Value.Count);
+            Assert.Equal(10, List.Value[0]);
+            Assert.Equal(1, Collection.Value.Count);
+            Assert.Equal(30, Collection.Value.ElementAt(0));
+            Assert.Equal(1, Set.Value.Count);
+            Assert.Equal(50, Set.Value.ElementAt(0));
+            Assert.Equal(1, Dictionary.Value.Keys.Count);
+            Assert.Equal("value1", Dictionary.Value["key1"]);
+        }
+    }
 }
