+++ src/libraries/System.Text.Json/tests/Serialization/TestClasses/TestClasses.NonGenericCollections.cs
@@ -432,4 +432,241 @@ namespace System.Text.Json.Serialization.Tests
     }
 
     public interface IDerivedIList : IList { }
+
+
+    public struct StructWrapperForIList : IList
+    {
+        private List<object> _list;
+
+        private void InitializeIfNull()
+        {
+            if (_list == null)
+            {
+                _list = new List<object>();
+            }
+        }
+
+        public object this[int index]
+        {
+            get
+            {
+                InitializeIfNull();
+                return _list[index];
+            }
+            set
+            {
+                InitializeIfNull();
+                _list[index] = value;
+            }
+        }
+
+        public bool IsFixedSize => ((IList)_list).IsFixedSize;
+
+        public bool IsReadOnly => false;
+
+        public int Count => _list == null ? 0 : _list.Count;
+
+        public bool IsSynchronized => ((IList)_list).IsSynchronized;
+
+        public object SyncRoot => ((IList)_list).SyncRoot;
+
+        public int Add(object value)
+        {
+            InitializeIfNull();
+            return ((IList)_list).Add(value);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            _list.Clear();
+        }
+
+        public bool Contains(object value)
+        {
+            InitializeIfNull();
+            return ((IList)_list).Contains(value);
+        }
+
+        public void CopyTo(Array array, int index)
+        {
+            InitializeIfNull();
+            ((IList)_list).CopyTo(array, index);
+        }
+
+        public IEnumerator GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((IList)_list).GetEnumerator();
+        }
+
+        public int IndexOf(object value)
+        {
+            InitializeIfNull();
+            return _list.IndexOf(value);
+        }
+
+        public void Insert(int index, object value)
+        {
+            InitializeIfNull();
+            _list.Insert(index, value);
+        }
+
+        public void Remove(object value)
+        {
+            InitializeIfNull();
+            _list.Remove(value);
+        }
+
+        public void RemoveAt(int index)
+        {
+            InitializeIfNull();
+            _list.RemoveAt(index);
+        }
+    }
+
+    public struct StructWrapperForIDictionary : IDictionary
+    {
+        private Dictionary<string, object> _dictionary;
+
+        private void InitializeIfNull()
+        {
+            if (_dictionary == null)
+            {
+                _dictionary = new Dictionary<string, object>();
+            }
+        }
+
+        public object this[object key]
+        {
+            get
+            {
+                InitializeIfNull();
+                return ((IDictionary)_dictionary)[key];
+            }
+            set
+            {
+                InitializeIfNull();
+                ((IDictionary)_dictionary)[key] = value;
+            }
+        }
+
+        public bool IsFixedSize => ((IDictionary)_dictionary).IsFixedSize;
+
+        public bool IsReadOnly => false;
+
+        public ICollection Keys => ((IDictionary)_dictionary).Keys;
+
+        public ICollection Values => ((IDictionary)_dictionary).Values;
+
+        public int Count => _dictionary.Count;
+
+        public bool IsSynchronized => ((IDictionary)_dictionary).IsSynchronized;
+
+        public object SyncRoot => ((IDictionary)_dictionary).SyncRoot;
+
+        public void Add(object key, object value)
+        {
+            InitializeIfNull();
+            ((IDictionary)_dictionary).Add(key, value);
+        }
+
+        public void Clear()
+        {
+            InitializeIfNull();
+            _dictionary.Clear();
+        }
+
+        public bool Contains(object key)
+        {
+            InitializeIfNull();
+            return ((IDictionary)_dictionary).Contains(key);
+        }
+
+        public void CopyTo(Array array, int index)
+        {
+            InitializeIfNull();
+            ((IDictionary)_dictionary).CopyTo(array, index);
+        }
+
+        public IDictionaryEnumerator GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((IDictionary)_dictionary).GetEnumerator();
+        }
+
+        public void Remove(object key)
+        {
+            InitializeIfNull();
+            ((IDictionary)_dictionary).Remove(key);
+        }
+
+        IEnumerator IEnumerable.GetEnumerator()
+        {
+            InitializeIfNull();
+            return ((IDictionary)_dictionary).GetEnumerator();
+        }
+    }
+
+    public class ClassWithStructIDictionaryWrapper
+    {
+        public StructWrapperForIDictionary Dictionary { get; set; }
+    }
+
+    public class ClassWithStructIListWrapper
+    {
+        public StructWrapperForIList List { get; set; }
+    }
+
+    public class SimpleTestClassWithStructCollectionWrappers : ITestClass
+    {
+        public StructWrapperForIList List { get; set; }
+        public StructWrapperForIDictionary Dictionary { get; set; }
+
+        public static readonly string s_json =
+            @"{" +
+            @"""List"" : [""Hello""]," +
+            @"""Dictionary"" : {""key1"" : ""value1""}" +
+            @"}";
+
+        public void Initialize()
+        {
+            List = new StructWrapperForIList() { "Hello" };
+            Dictionary = new StructWrapperForIDictionary { { "key1", "value1" } };
+        }
+
+        public void Verify()
+        {
+            Assert.Equal(1, List.Count);
+            Assert.Equal("Hello", ((JsonElement)List[0]).GetString());
+            Assert.Equal(1, Dictionary.Keys.Count);
+            Assert.Equal("value1", ((JsonElement)Dictionary["key1"]).GetString());
+        }
+    }
+
+    public struct SimpleTestStructWithNullableStructCollectionWrappers : ITestClass
+    {
+        public StructWrapperForIList? List { get; set; }
+        public StructWrapperForIDictionary? Dictionary { get; set; }
+
+        public static readonly string s_json =
+            @"{" +
+            @"""List"" : [""Hello""]," +
+            @"""Dictionary"" : {""key1"" : ""value1""}" +
+            @"}";
+
+        public void Initialize()
+        {
+            List = new StructWrapperForIList() { "Hello" };
+            Dictionary = new StructWrapperForIDictionary { { "key1", "value1" } };
+        }
+
+        public void Verify()
+        {
+            Assert.Equal(1, List.Value.Count);
+            Assert.Equal("Hello", ((JsonElement)List.Value[0]).GetString());
+            Assert.Equal(1, Dictionary.Value.Keys.Count);
+            Assert.Equal("value1", ((JsonElement)Dictionary.Value["key1"]).GetString());
+        }
+    }
 }
