+++ src/libraries/System.Text.Json/tests/System.Text.Json.Tests.csproj
@@ -117,10 +117,8 @@
     <Compile Include="Serialization\TestClasses\TestClasses.Constructor.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.GenericCollections.cs" />
-    <Compile Include="Serialization\TestClasses\TestClasses.GenericCollectionsStruct.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.ImmutableCollections.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.NonGenericCollections.cs" />
-    <Compile Include="Serialization\TestClasses\TestClasses.NonGenericCollectionsStruct.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.Polymorphic.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.SimpleTestClass.cs" />
     <Compile Include="Serialization\TestClasses\TestClasses.SimpleTestClassWithFields.cs" />
